requirements.txt
Flask-OAuthlib==0.9
Flask-Testing==0.4.2
Flask-Script==0.6.2
requests==2.5.3 # testonly
nose==1.3.4 # testonly